Reduce number of Regex::is_match calls
Regex::is_match accounts for a large portion of time spent parsing
Markdown text. This seems to be because the regexes are checked for
matches iteratively with varying size slices in
markdown::parser::span::parse_spans. Code, emphasis, and strong checks
were double the time of the other span checks due to using two regexes
and calling is_match on both of them with each iteration.

This commit combines the regexes in code, emphasis, and strong tag
parsing into a single regex for each tag. The result is that checks for
these tags now take roughly the same amount of time as for the checks
for other tags. The Regex::is_match calls still dominate time spent in
parsing but this is a marginal improvement.

I've also added a very basic benchmark based on the "easy.txt" test
fixture. This is certainly not a representative markdown document but
running markdown::to_html against it does show improvement with this
change. This seems to be because it uses both emphasis and strong tags
and those are checked after code tags in parse_span, therefore all three
of the improved tag checks are exercised in this benchmark.